pyproject.toml
[project]
name = "topic4-arag-eval"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.5",
    "gradio>=5.0",
    "httpx>=0.27",
    "openai>=1.0",
    "pandas>=3.0.2",
    "pypdf>=5.0",
    "python-docx>=1.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_rag"]
